+++ tests/Data/Stubs/FromUsersQueryExportWithMapping.php
@@ -2,6 +2,8 @@
 
 namespace Maatwebsite\Excel\Tests\Data\Stubs;
 
+use Illuminate\Database\Eloquent\Builder as EloquentBuilder;
+use Illuminate\Database\Eloquent\Relations\Relation;
 use Illuminate\Database\Query\Builder;
 use Maatwebsite\Excel\Concerns\Exportable;
 use Maatwebsite\Excel\Concerns\FromQuery;
@@ -15,7 +17,7 @@ class FromUsersQueryExportWithMapping implements FromQuery, WithMapping, WithEve
     use Exportable;
 
     /**
-     * @return Builder
+     * @return Builder|EloquentBuilder|Relation
      */
     public function query()
     {