+++ CHANGELOG.md
@@ -4,6 +4,8 @@ All notable changes to this project will be documented in this file.
 
 ## [Unreleased]
 
+- Fix return type of `FromQuery::query()`
+
 ## [3.1.35] - 2022-01-04
 
 ### Fixed